revert: Revert prosody_features to int64 data type
commit 3340825でfloat32に変更したprosody_featuresをint64に戻す。
float32変更はONNX推論で不自然な音声を生成する問題を引き起こしていた。

## 問題
- ONNX float32推論: 不自然な音声（話速・音質・イントネーション異常）
- PyTorch推論（int64 → .float()変換）: 正常な音声

## 根本原因
- **float32入力**: `.float()`がno-opとなり、ONNXグラフにCastノードが含まれない/最適化される
- **int64入力**: `.float()`が明示的なCastノードを生成し、正しい動作
- グラフ構造の違いがDuration Predictorの数値計算に影響

models.py Line 654:
```python
prosody_proj = self.prosody_proj(prosody_features.float())
```

## 検証結果
長文3種類（10-14秒）でテスト:

| テキスト | PyTorch | ONNX float32 | ONNX int64 | 評価 |
|---------|---------|--------------|-----------|------|
| 説明文 | 625KB (14.50秒) | 563KB (13.06秒) | 615KB (14.26秒) | PyTorch正常、float32異常、int64正常 |
| 物語風 | 607KB (14.07秒) | 548KB (12.70秒) | 621KB (14.41秒) | PyTorch正常、float32異常、int64正常 |
| 技術説明 | 544KB (12.61秒) | 565KB (13.10秒) | 585KB (13.56秒) | PyTorch正常、float32異常、int64正常 |

**ユーザー評価**:
- PyTorch: "全て問題なし" ✅
- ONNX float32: "ダメ" ❌
- ONNX int64 (修正後): "問題ないです" ✅

## 変更内容

### Python (4ファイル)
1. **export_onnx.py** (Line 225):
   - `torch.zeros(..., dtype=torch.float32)` → `torch.zeros(..., dtype=torch.long)`

2. **infer_onnx.py** (Lines 100, 102, 104, 108):
   - `[0.0, 0.0, 0.0]` → `[0, 0, 0]`
   - `float(pf["a1"])` → `pf["a1"]`
   - `dtype=np.float32` → `dtype=np.int64`

3. **test_pytorch_onnx_parity.py** (Lines 46, 54, 158, 187, 267-268):
   - docstring: "float32" → "int64"
   - `dtype=np.float32` → `dtype=np.int64`
   - assertion: `tensor(float)` → `tensor(int64)`

4. **conftest.py** (Line 105):
   - `dtype=torch.float32` → `dtype=torch.long`

### C++ (1ファイル)
5. **piper.cpp** (複数箇所):
   - `std::vector<float>` → `std::vector<int64_t>`
   - `CreateTensor<float>` → `CreateTensor<int64_t>`
   - `static_cast<float>(pf.a1)` → `pf.a1`（直接代入）

## テスト結果
```bash
pytest src/python/tests/test_pytorch_onnx_parity.py -v
7 passed, 10 warnings ✅
```

## 意味論的な正しさ
A1/A2/A3は整数値:
- **A1**: -4, -3, ..., 0, 1, ... (アクセント核からの相対位置)
- **A2**: 1, 2, 3, ... (モーラ位置)
- **A3**: 1-10+ (総モーラ数)

**int64が意味論的に正しい型**。`.float()`による変換はモデル内部で明示的に行われるべき。

+++ src/python/piper_train/infer_onnx.py
@@ -97,15 +97,15 @@ def main():
                 prosody_array = []
                 for pf in prosody_features_data:
                     if pf is None:
-                        prosody_array.append([0.0, 0.0, 0.0])
+                        prosody_array.append([0, 0, 0])
                     else:
-                        prosody_array.append([float(pf["a1"]), float(pf["a2"]), float(pf["a3"])])
+                        prosody_array.append([pf["a1"], pf["a2"], pf["a3"]])
                 prosody_features = np.expand_dims(
-                    np.array(prosody_array, dtype=np.float32), 0
+                    np.array(prosody_array, dtype=np.int64), 0
                 )
             else:
-                # No prosody data provided - use zeros (float32)
-                prosody_features = np.zeros((1, text.shape[1], 3), dtype=np.float32)
+                # No prosody data provided - use zeros (int64)
+                prosody_features = np.zeros((1, text.shape[1], 3), dtype=np.int64)
             inputs["prosody_features"] = prosody_features
 
         start_time = time.perf_counter()
